Improved filament load and unload sequence
to make load faster and unload too. Also focused on unload to add a quick ramming so that the filament gets pulled out quickly and if possible without a long strand. Works in estiamted 95% of unloads without long strands.

Tried with Prusa PLA and ABS-T (Easy ABS) (both Filament PM)